+++ demo/src/main/java/com/teleconsulting/demo/service/PatientServiceImpl.java
@@ -1,5 +1,6 @@
 package com.teleconsulting.demo.service;
 
+import com.teleconsulting.demo.handler.NotFoundException;
 import com.teleconsulting.demo.model.Doctor;
 import com.teleconsulting.demo.model.Patient;
 import com.teleconsulting.demo.repository.PatientRepository;
@@ -25,21 +26,33 @@ public class PatientServiceImpl implements PatientService{
         if(pat.isPresent()){
             return pat.get();
         }
-        throw new RuntimeException("Doctor is not found for the id :"+id);
+        throw new NotFoundException("Doctor is not found for the id :"+id);
     }
 
     @Override
     public void deletePatient(Long id) {
+        Optional<Patient> existingPatient = patientRepository.findById(id);
+        if (existingPatient.isEmpty()) {
+            throw new NotFoundException("Patient doesn't exists !!\n");
+        }
         patientRepository.deleteById(id);
     }
 
     @Override
     public Patient updatePatient(Patient patient) {
+        Optional<Patient> existingPatient = patientRepository.findById(patient.getId());
+        if (existingPatient.isEmpty()) {
+            throw new NotFoundException("Patient doesn't exists !!\n");
+        }
         return patientRepository.save(patient);
     }
 
     @Override
     public Patient savePatient(Patient patient) {
+        Optional<Patient> existingPatient = patientRepository.findByPhoneNumber(patient.getPhoneNumber());
+        if (existingPatient.isPresent()) {
+            throw new NotFoundException("Phone Number already exists: " + patient.getPhoneNumber());
+        }
         return patientRepository.save(patient);
     }
 }
